make only acutal text-part of (short) PopupIcons clickable
I see that icons must have a fixed width in order to center them. But it resulted in the whole area being clickable (even the transparent areas right and left of the text). This behaviour is especially confusing if texts are short, but right next to each other.

This uses [pointer-events](http://caniuse.com/pointer-events) to make the transparent areas right and left of PopupIcons non-clickable. And it doesn't harm older browsers.